requirements.txt
astroid==3.1.0
blinker==1.7.0
cfgv==3.4.0
click==8.1.7
dill==0.3.8
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
identify==2.5.35
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
nodeenv==1.8.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
pylint==3.1.0
PyYAML==6.0.1
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
virtualenv==20.25.1
Werkzeug==3.0.1
fastai==2.7.14